+++ src/datatable/js/highlight.js
@@ -3,7 +3,7 @@
 
  @module datatable
  @submodule datatable-highlight
- @since @SINCE@
+ @since 3.13.0
  */
 
 
@@ -11,7 +11,7 @@ var getClassName = Y.ClassNameManager.getClassName;
 
 /**
  @class DataTable.Highlight
- @since @SINCE@
+ @since 3.13.0
  */
 function Highlight() {}
 
@@ -22,7 +22,7 @@ Highlight.ATTRS = {
 
      @attribute highlightRows
      @default false
-     @since @SINCE@
+     @since 3.13.0
      */
     highlightRows: {
         value: false,
@@ -36,7 +36,7 @@ Highlight.ATTRS = {
 
      @attribute highlightCols
      @default false
-     @since @SINCE@
+     @since 3.13.0
      */
     highlightCols: {
         value: false,
@@ -50,7 +50,7 @@ Highlight.ATTRS = {
 
      @attribute highlightCells
      @default false
-     @since @SINCE@
+     @since 3.13.0
      */
     highlightCells: {
         value: false,
@@ -70,7 +70,7 @@ Highlight.prototype = {
      @public
      @property highlightClassNames
      @type Object
-     @since @SINCE@
+     @since 3.13.0
      */
     highlightClassNames: {
         row: getClassName(NAME, 'row'),
@@ -87,7 +87,7 @@ Highlight.prototype = {
      @protected
      @property _colSelector
      @type String
-     @since @SINCE@
+     @since 3.13.0
      */
     _colSelector: '.{prefix}-data .{prefix}-col-{col}',
 
@@ -99,7 +99,7 @@ Highlight.prototype = {
      @protected
      @property _colNameRegex
      @type String
-     @since @SINCE@
+     @since 3.13.0
      */
     _colNameRegex: '{prefix}-col-(\\S*)',
 
@@ -110,7 +110,7 @@ Highlight.prototype = {
      @protected
      @property _highlightDelegates
      @type Object
-     @since @SINCE@
+     @since 3.13.0
      */
     _highlightDelegates: {},
 
@@ -124,7 +124,7 @@ Highlight.prototype = {
      @method _setHighlightRows
      @param {Boolean} val
      @return val
-     @since @SINCE@
+     @since 3.13.0
      */
     _setHighlightRows: function (val) {
         var del = this._highlightDelegates;
@@ -153,7 +153,7 @@ Highlight.prototype = {
      @method _setHighlightCols
      @param {Boolean} val
      @return val
-     @since @SINCE@
+     @since 3.13.0
      */
     _setHighlightCols: function (val) {
         var del = this._highlightDelegates;
@@ -182,7 +182,7 @@ Highlight.prototype = {
      @method _setHighlightCells
      @param {Boolean} val
      @return val
-     @since @SINCE@
+     @since 3.13.0
      */
     _setHighlightCells: function (val) {
         var del = this._highlightDelegates;
@@ -211,7 +211,7 @@ Highlight.prototype = {
      @protected
      @method _highlightRow
      @param {EventFacade} e Event from the hover event
-     @since @SINCE@
+     @since 3.13.0
      */
     _highlightRow: function (e) {
         e.currentTarget.toggleClass(this.highlightClassNames.row, (e.phase === 'over'));
@@ -226,7 +226,7 @@ Highlight.prototype = {
      @protected
      @method _highlightCol
      @param {EventFacade} e Event from the hover event
-     @since @SINCE@
+     @since 3.13.0
      */
     _highlightCol: function(e) {
         var colName = this._colNameRegex.exec(e.currentTarget.getAttribute('class')),
@@ -247,7 +247,7 @@ Highlight.prototype = {
      @protected
      @method _highlightCell
      @param {EventFacade} e Event from the hover event
-     @since @SINCE@
+     @since 3.13.0
      */
     _highlightCell: function(e) {
         e.currentTarget.toggleClass(this.highlightClassNames.cell, (e.phase === 'over'));
@@ -260,7 +260,7 @@ Highlight.prototype = {
 
      @protected
      @method _buildColSelRegex
-     @since @SINCE@
+     @since 3.13.0
      */
     _buildColSelRegex: function () {
         var str = this._colNameRegex,